Move confirmation tests to their own test class
